refactor CO2 and wind sensor names
