+++ sensor/src/hardware/__init__.py
@@ -5,12 +5,12 @@ import filelock
 
 from src import custom_types, utils
 from .bme280_sensor import BME280SensorInterface
-from .co2_sensor import CO2SensorInterface
+from .gmp343_sensor import CO2SensorInterface
 from .pump import PumpInterface
 from .sht45_sensor import SHT45SensorInterface
 from .ups import UPSInterface
 from .valves import ValveInterface
-from .wind_sensor import WindSensorInterface
+from .wxt532_sensor import WindSensorInterface
 
 
 class HwLock(TypedDict):
