+++ sensor/src/hardware/wxt532_sensor.py
